backend/requirements.txt
fastapi
uvicorn
python-docx
pdfplumber
ollama
python-multipart
json-repair